requirements.txt
pandas
requests
bs4
tabulate
python-dotenv